rename simplenet naivenet
